# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Talos API
  description: >
    Source of truth for the talos-api public (/api/v1) and internal (/internal/v1) REST surface.
    The Angular client is generated from this file (openapi-generator, typescript-angular); CI
    diffs the running API's springdoc output against it. Endpoints are added phase by phase per
    Section 10 of the implementation plan — this file currently covers Phase 1 (auth),
    Phase 3 (project registry), Phase 4 (task board), Phase 5 (agent run lifecycle, API side),
    Phase 6 (cancel + internal changes ingest), Phase 8 (review/approval flow: diff serving and
    the approvals endpoints), Phase 9 (git push and PR workflow: integrations, the
    git-token/pull-request internal endpoints, and GET /runs/{id}/pull-request), Phase 10
    (Dokploy integration), Phase 12 (chat trigger support), Phase 13 (memory ingestion and
    internal retrieval), Phase 14 (per-run cost tracking and advisory recommendations), and
    Phase 15 (multi-user RBAC enforcement and user management).
  version: "0.15.0"
servers:
  - url: /api/v1
    description: "Public API (JWT via Authorization: Bearer)"
  - url: /internal/v1
    description: "Internal API (service token via X-Talos-Internal-Token)"
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: List users (Section 16 Phase 15). OWNER only.
      tags: [users]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Error"
    post:
      operationId: createUser
      summary: >-
        Create a user with an operator-set password (not a token-based email invite -- Talos has no
        mail infrastructure yet). OWNER only.
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          description: A user with this email already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{id}:
    patch:
      operationId: updateUser
      summary: >-
        Change a user's role and/or active status. OWNER only; an OWNER cannot change their own
        role or deactivate themselves through this endpoint (use another OWNER account).
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          description: Attempted to change one's own role or deactivate oneself (CANNOT_MODIFY_OWN_ACCOUNT).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /projects/{id}/memory/documents:
    post:
      operationId: ingestProjectMemoryDocument
      summary: >
        Phase 13: ingest an operator note or context document into project-scoped memory. The API
        owns chunking, masking, embedding, and persistence; no Python service writes memory rows.
      tags: [memory]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryDocumentRequest"
      responses:
        "201":
          description: Created, or the existing identical document if already ingested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryDocument"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
...
  /integrations:
    get:
      operationId: listIntegrations
      summary: List configured integrations (Section 10.2). Secrets are never returned. OWNER only (Section 16 Phase 15).
      tags: [integrations]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Integration"
        "403":
          $ref: "#/components/responses/Error"
    post:
      operationId: createIntegration
      summary: >-
        Configure an integration (e.g. a GitHub PAT) -- the secret is encrypted and stored behind a
        secret_ref (Section 12.2). OWNER only (Section 16 Phase 15).
      tags: [integrations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Integration"
        "403":
          $ref: "#/components/responses/Error"
  /integrations/{id}/test:
    post:
      operationId: testIntegration
      summary: Verify the stored credential works (e.g. calls GET /user against GitHub). OWNER only (Section 16 Phase 15).
      tags: [integrations]
      parameters:
        - $ref: "#/components/parameters/IntegrationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestIntegrationResponse"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /approvals:
    get:
      operationId: listApprovals
      summary: List approvals (Section 10.2).
      tags: [approvals]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ApprovalStatus"
        - name: runId
          in: query
          required: false
          description: Phase 8 addition -- lets the Review page (/review/:runId) find "the approval for this run".
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          required: false
          description: >
            Phase 10 addition -- filters by approvalType ('RUN_RESULT'|'DEPLOY'). A COMPLETED run
            can have both a decided RUN_RESULT approval and a fresh PENDING DEPLOY approval under
            the same runId, so callers that want just one kind need this.
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageApproval"
  /approvals/{id}:
    get:
      operationId: getApproval
      summary: Get an approval, its run, and the diff summary.
      tags: [approvals]
      parameters:
        - $ref: "#/components/parameters/ApprovalId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDetail"
        "404":
          $ref: "#/components/responses/Error"
  /approvals/{id}/approve:
    post:
      operationId: approveApproval
      summary: >-
        Approve a PENDING approval (Section 8.2's only path to WAITING_APPROVAL -> APPROVED).
        REVIEWER+ only. The user who requested the underlying run/deploy cannot approve it
        themselves (403 SELF_APPROVAL_FORBIDDEN) unless they are OWNER, in which case the override
        is audited (Section 16 Phase 15).
      tags: [approvals]
      parameters:
        - $ref: "#/components/parameters/ApprovalId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "403":
          description: Role below REVIEWER, or a non-OWNER self-approval attempt (SELF_APPROVAL_FORBIDDEN).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          description: Approval is already decided.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /approvals/{id}/reject:
    post:
      operationId: rejectApproval
      summary: >-
        Reject a PENDING approval; the run moves to REJECTED and the task returns to READY.
        REVIEWER+ only; same self-approval prohibition as approve (Section 16 Phase 15).
      tags: [approvals]
      parameters:
        - $ref: "#/components/parameters/ApprovalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          description: Approval is already decided.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /approvals/{id}/request-changes:
    post:
      operationId: requestApprovalChanges
      summary: >
        Request changes on a PENDING approval. Section 8.2 has no run status for "changes
        requested" -- the approval keeps this distinct outcome, but the run is driven through the
        REJECTED edge so the task returns to READY for rework. REVIEWER+ only; same self-approval
        prohibition as approve (Section 16 Phase 15).
      tags: [approvals]
      parameters:
        - $ref: "#/components/parameters/ApprovalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestChangesRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          description: Approval is already decided.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /internal/v1/projects/{id}/memory/documents:
    post:
      operationId: internalIngestProjectMemoryDocument
      summary: >
        Phase 13: ingest context.docs content discovered by the orchestrator. The API still owns
        masking, chunking, embedding, and persistence.
      tags: [internal]
      security:
        - serviceToken: []
      servers:
        - url: /internal/v1
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryDocumentRequest"
      responses:
        "201":
          description: Created, or the existing identical document if already ingested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryDocument"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceToken:
      type: apiKey
      in: header
      name: X-Talos-Internal-Token
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ApprovalId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IntegrationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Error envelope (Section 10.1).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: Error envelope returned on every non-2xx response (Section 10.1).
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: SCREAMING_SNAKE error code enumerated in this spec.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    Role:
      type: string
      description: >-
        Section 9.3/Phase 15: a strict hierarchy, OWNER > MAINTAINER > REVIEWER > VIEWER -- each
        role implies every permission of the ones after it.
      enum: [OWNER, MAINTAINER, REVIEWER, VIEWER]
    User:
      type: object
      required: [id, email, name, role, active, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        active:
          type: boolean
          description: A deactivated user can no longer log in; already-issued JWTs remain valid until their 24h expiry.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [email, name, password, role]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
          minLength: 8
        role:
          $ref: "#/components/schemas/Role"
    UpdateUserRequest:
      type: object
      description: Only non-null fields are applied.
      properties:
        role:
          $ref: "#/components/schemas/Role"
        active:
          type: boolean
...
    MemorySourceType:
      type: string
      enum: [RUN_SUMMARY, OPERATOR_NOTE, CONTEXT_DOC]
    MemoryDocumentRequest:
      type: object
      required: [sourceType, content]
      properties:
        sourceType:
          $ref: "#/components/schemas/MemorySourceType"
        sourceRef:
          type: string
          description: Operator-controlled reference such as a doc path, note id, or external URL.
        title:
          type: string
        content:
          type: string
          description: Plaintext content; talos-api masks common secret shapes before persistence.
    MemoryDocument:
      type: object
      required: [id, projectId, sourceType, sourceRef, title, createdAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        sourceType:
          $ref: "#/components/schemas/MemorySourceType"
        sourceRef:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
...
    RunStatus:
      type: string
      enum: [CREATED, QUEUED, PREPARING_WORKSPACE, RUNNING_AGENT, RUNNING_TESTS, REVIEWING,
        WAITING_APPROVAL, APPROVED, REJECTED, COMPLETED, FAILED, CANCELLED]
    TestStatus:
      type: string
      enum: [NOT_RUN, PASSED, FAILED, ERROR]
    ReviewStatus:
      type: string
      enum: [CLEAN, RISK_FLAGGED]
...
    Run:
      type: object
      required: [id, taskId, projectId, status, agentKey, providerAuthMode, testStatus, reviewStatus,
        createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/RunStatus"
        agentKey:
          type: string
        providerAuthMode:
          type: string
        prompt:
          type: string
        branchName:
          type: string
        workspacePath:
          type: string
        summary:
          type: string
        testStatus:
          $ref: "#/components/schemas/TestStatus"
        reviewStatus:
          $ref: "#/components/schemas/ReviewStatus"
        errorMessage:
          type: string
        exitCode:
          type: integer
        inputTokens:
          type: integer
          description: Phase 14 (Section 16) normalized usage metadata; null when the adapter can't report it.
        outputTokens:
          type: integer
        costUsd:
          type: number
          description: >-
            Provider-reported price in USD. Always null for provider_auth_mode=subscription_local
            (Section 13) -- never a fabricated/estimated price.
        costModel:
          type: string
        timeoutAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    GitChangeType:
      type: string
      enum: [ADDED, MODIFIED, DELETED, RENAMED]
    GitChange:
      type: object
      required: [filePath, changeType, additions, deletions]
      properties:
        filePath:
          type: string
        changeType:
          $ref: "#/components/schemas/GitChangeType"
        additions:
          type: integer
        deletions:
          type: integer
    GitChangeSummary:
      description: >
        Output-only counterpart to GitChange (Phase 8): adds the Section 12.1 post-run policy scan
        result, shown in the Review Center's file list and risk panel.
      allOf:
        - $ref: "#/components/schemas/GitChange"
        - type: object
          required: [riskFlagged]
          properties:
            riskFlagged:
              type: boolean
            matchedPattern:
              type: string
              description: The policy.yaml/talos.yaml pattern that flagged this file, if any.
...
    ApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, CHANGES_REQUESTED, EXPIRED]
    Approval:
      type: object
      required: [id, taskId, runId, approvalType, requestedAction, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        approvalType:
          type: string
        requestedAction:
          type: string
        status:
          $ref: "#/components/schemas/ApprovalStatus"
        requestedBy:
          type: string
          format: uuid
        approvedBy:
          type: string
          format: uuid
        approvedAt:
          type: string
          format: date-time
        notes:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        environment:
          type: string
          description: Phase 10 -- set only on DEPLOY-type approvals; which environment this approval gates.
    ApprovalDetail:
      type: object
      required: [approval, run, changes]
      properties:
        approval:
          $ref: "#/components/schemas/Approval"
        run:
          $ref: "#/components/schemas/Run"
        changes:
          type: array
          items:
            $ref: "#/components/schemas/GitChangeSummary"
    ApproveRequest:
      type: object
      properties:
        notes:
          type: string
    RejectRequest:
      type: object
      required: [notes]
      properties:
        notes:
          type: string
    RequestChangesRequest:
      type: object
      required: [notes]
      properties:
        notes:
          type: string
    PageApproval:
      type: object
      required: [content, page, size, totalElements]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Approval"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
...
    Integration:
      type: object
      description: Section 10.2 -- secrets are never included (they live behind SecretValue, referenced only by secret_ref).
      required: [id, type, name, enabled, createdAt]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          description: "'github', 'dokploy', 'anthropic', ..."
        name:
          type: string
        configJson:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    IntegrationCreateRequest:
      type: object
      required: [type, name]
      properties:
        type:
          type: string
        name:
          type: string
        configJson:
          type: object
          additionalProperties: true
        secret:
          type: string
          description: Plaintext credential (e.g. a GitHub fine-grained PAT); encrypted at rest and never echoed back.
        authMode:
          type: string
          description: "'pat', 'deploy_key', ... (integration_credentials.auth_mode, Section 9.2)"
    TestIntegrationResponse:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
        message:
          type: string
...